+++ poem/CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+# [1.2.37] 2022-1-18
+
+- Add `EndpointExt::to_response` method.
+
 # [1.2.34] 2022-1-14
 
 - Add `TcpAcceptor::from_std` and `UnixAcceptor::from_std` methods.